backend/requirements.txt
asgiref==3.7.2
Django==3.2.16
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flake8==7.0.0
flake8-broken-line==1.0.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
isort==5.13.2
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
pep8-naming==0.13.3
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
